fix: rename routes.tsx to AppRoutes.tsx
